requirements.txt
apache-airflow==2.2.3
pandas==1.3.3
mysql-connector-python==8.0.26
pytest==6.2.5
Django==3.2.3
psycopg2-binary==2.9.6
